+++ src/main/kotlin/org/bibletranslationtools/fetcher/impl/repository/PortGatewayLanguageCatalog.kt
@@ -1,4 +1,4 @@
-package org.bibletranslationtools.fetcher.repository.implementations
+package org.bibletranslationtools.fetcher.impl.repository
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
